tests/wpt/tests/html/interaction/focus/focusgroup/tentative/sequential-navigation/directional-input-handler-tab-escape.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Native directional input handler elements and Tab escape</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div id=before tabindex=0>Before focusgroup</div>

<div id=toolbar focusgroup="toolbar nomemory">
  <button tabindex=0 id=bold type="button">Bold</button>
  <!-- Native text input is a native directional input handler for both axes. -->
  <input id=search type="text" value="Search" />
  <button tabindex=0 id=italic type="button">Italic</button>
</div>

<div id=after tabindex=0>After focusgroup</div>

<script>
  // Directional input CAN navigate TO a native directional input handler (entry works).
  promise_test(async t => {
    const bold = document.getElementById("bold");
    const search = document.getElementById("search");

    await focusAndWait(bold);
    assert_equals(document.activeElement, bold, "Precondition: focus starts on bold button");

    await focusAndSendDirectionalInput(bold, kRight);
    assert_equals(document.activeElement, search,
                  "Right should navigate TO native directional input handler (entry allowed)");
  }, "Directional input can navigate TO native directional input handler (entry)");

  // Directional input CANNOT navigate FROM a native directional input handler (exit blocked).
  // The element's native directional input behavior takes precedence.
  promise_test(async t => {
    const search = document.getElementById("search");

    await focusAndWait(search);
    assert_equals(document.activeElement, search, "Precondition: focus starts on native directional input handler");

    await focusAndSendDirectionalInput(search, kRight);
    assert_equals(document.activeElement, search,
                  "Right should not move focus out of native directional input handler (native behavior)");

    await focusAndSendDirectionalInput(search, kLeft);
    assert_equals(document.activeElement, search,
                  "Left should not move focus out of native directional input handler (native behavior)");
  }, "Directional input cannot navigate FROM native directional input handler (exit blocked)");

  // Tab from native directional input handler should move to the next adjacent
  // focusgroup item.
  promise_test(async t => {
    const search = document.getElementById("search");
    const italic = document.getElementById("italic");

    await focusAndWait(search);
    assert_equals(document.activeElement, search, "Precondition: focus starts on native directional input handler");

    await sendTabForward();
    assert_equals(document.activeElement, italic,
                  "Tab from native directional input handler should move to next adjacent focusgroup item");
  }, "Tab from native directional input handler moves to next adjacent item");

  // Shift+Tab from native directional input handler: focus moves to the previous
  // adjacent focusgroup item.
  promise_test(async t => {
    const bold = document.getElementById("bold");
    const search = document.getElementById("search");

    await focusAndWait(search);
    assert_equals(document.activeElement, search, "Precondition: focus starts on native directional input handler");

    await navigateFocusBackward();
    assert_equals(document.activeElement, bold,
                  "Shift+Tab from native directional input handler should go to previous adjacent focusgroup item");
  }, "Shift+Tab from native directional input handler goes to previous adjacent item");
</script>
